// backend/src/modules/storage/enums/upload-intent.enum.ts
export enum UploadIntent {
  PROFILE_AVATAR = 'PROFILE_AVATAR',
  PROFILE_BANNER = 'PROFILE_BANNER',
  COMMUNITY_BANNER = 'COMMUNITY_BANNER',
  COMMUNITY_THUMBNAIL = 'COMMUNITY_THUMBNAIL',
  COMMUNITY_COVER = 'COMMUNITY_COVER',
  FEED_IMAGE = 'FEED_IMAGE',
  EXPERIENCE_IMAGE = 'EXPERIENCE_IMAGE',
  PROVIDER_IMAGE = 'PROVIDER_IMAGE',
  DOCUMENT = 'DOCUMENT',
  CERTIFICATE = 'CERTIFICATE',
}